feat: Introduce an add-on interface
* Add manager export and a preview module
* Adapt the example.
* Adapt the documentation

The modification of preview.js is still needed, unfortunately.
