+++ CHANGELOG.md
@@ -1,14 +1,18 @@
-# [2.0.0](https://github.com/prantlf/storybook-multilevel-sort/compare/v1.2.1...v2.0.0) (2023-05-14)
-
+# [2.0.0-next.2](https://github.com/prantlf/storybook-multilevel-sort/compare/v2.0.0-next.1...v2.0.0-next.2) (2023-05-14)
 
 ### Features
 
-* Support Storybook 7 and docs pages ([#10](https://github.com/prantlf/storybook-multilevel-sort/issues/10)) ([1065337](https://github.com/prantlf/storybook-multilevel-sort/commit/10653377d94d46763e3c0885bc2386bf42a5d8d2))
+* Describe how to modify preview.js maually, unfortunately ([9695988](https://github.com/prantlf/storybook-multilevel-sort/commit/9695988cdef4e6b5894c7f9d32d61b90a9dbc060))
 
+# [2.0.0-next.1](https://github.com/prantlf/storybook-multilevel-sort/compare/v2.0.0-next.0...v2.0.0-next.1) (2023-05-08)
 
-### BREAKING CHANGES
+### Bug Fixes
 
-* Only Storybook 7 is supported. If you use Storybook 6 or older, stay with the versions 1.x. Also, the minimum supported version of Node.js is 16. Except for the version requirements, the configuration of the custom sorting needs to be specified by a different interface. See the documentation about migration for more information.
+* Make sortOrder option not mandatory, add tests ([ff2417a](https://github.com/prantlf/storybook-multilevel-sort/commit/ff2417a473de31cda9fe258c0534006ca22f25c5))
+
+### Features
+
+* Introduce an add-on interface ([4bdbd77](https://github.com/prantlf/storybook-multilevel-sort/commit/4bdbd77b8cdaf77bc6f59f9e3a0566d511eaaa7e))
 
 # [2.0.0-next.0](https://github.com/prantlf/storybook-multilevel-sort/compare/v1.2.0...v2.0.0-next.0) (2023-05-07)
