+++ LICENSE
@@ -32,5 +32,3 @@ Our license allows you to use the content:
 * For an unlimited number of times without interruption
 * Anywhere in the world
 * To make changes and create derivative works
-
-See also docs/icon-license.pdf.
