+++ lib/preview.js
@@ -1,0 +1,8 @@
+// export default {
+//   parameters: {
+//     options: {
+//       storySort: (story1, story2) =>
+//         globalThis['storybook-multilevel-sort:storySort'](story1, story2)
+//     }
+//   }
+// }
